Represent sizes as longs in Partition

+++ app/src/main/java/eu/depau/etchdroid/utils/PartitionBuilder.kt
@@ -6,7 +6,7 @@ import eu.depau.etchdroid.enums.PartitionType
 class PartitionBuilder {
     var number: Int? = null
     var offset: Int? = null
-    var size: Int? = null
+    var size: Long? = null
     var partType: PartitionType? = null
     var partLabel: String? = null
     var fsType: FilesystemType? = null
